# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация",
        "title": "ВИМБ API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/mq/queues": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Состояние очередей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MQ"
                ],
                "summary": "Состояние очередей.",
                "operationId": "routes-get-mq-queue-metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.MqInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
        "/api/v1/spot/set-position": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет позиционирование спота..",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Выполняет позиционирование спота..",
                "operationId": "routes-set-spot-position",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerSetSpotPositionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.StreamResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "request": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerSetSpotPositionRequest": {
            "type": "object",
            "properties": {
                "Distance": {
                    "type": "string"
                },
                "SpotID": {
                    "type": "string"
                }
            }
        },
...
        "routes.MqInfo": {
            "type": "object",
            "properties": {
                "consumers": {
                    "type": "integer"
                },
                "messages": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}